Double the extra native heap allowance.
This is in response to:
```
[2020-02-16 18:42:22: INFO] Finished 'renaissance__movie-lens(0)' (default-ext variant) under 'openj9'
[2020-02-16 18:42:25: WARNING] New dmesg lines:
[ 1483.432591] mmap: JIT Compilation (1097): VmData 15033163776 exceed data ulimit 15032385536. Update limits or use boot option ignore_rlimit_data.
```

I was going to skip this benchmark on OpenJ9 (since the other VMs don't
crash like this), but I decided to give it a second chance. After all,
we only gave the VM 2GiB (excluding the Java heap) for it's own use.

If this crashes again, I think we will skip this benchmark though.